chore(riklet): remove dead code and cleanup
* Removed unused tap related commands
* Create net_utils file to have some utilitary commands
* Remove net configuration and error from previous setup
